Shifted archives and resources to the right of "select a design."
See the diffs for details. What the diffs won't tell you is that the reason the float for the archives wasn't taking effect is that the width of the parent element was too narrow. In other words, if we pretend 'select a design' was 501 pixels, and 'archives' was 200 pixels, if the containing element was only 700 pixels, 'archives' would not be narrow enough to float up. In this case, I eliminated the padding (some of which you may wish to add back). If desired, I can make a diagram to illustrate the problem.